refactor(notification-settings-screen-ui): PR changes
